Refactor dealing to return actions instead of modifying state
This is part of a larger refactor to push all the handling for modifying
state into a single place. The idea is that actions that would modify
state will return a sequence of actions that should be applied to state
instead. Eventually, a single handler will process all the actions and
make the necessary state modifications. This will isolate state changes
to a single location and allow for better error handling.